test: extract dynamic filter tests into a child module
